archive/ported_reference_agents/reference-agno-agents/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
agno>=0.2.0
google-genai
python-dotenv>=1.0.0
requests>=2.31.0
# Native Tools Dependencies
duckduckgo-search
newspaper4k
lxml_html_clean
# Local Khala package
../../packages/khala-agentmemory
Pillow
pyyaml
